Fixed input and edge of map issue
Previously, the return key always added a line to text area even if I
removed all value from the textarea afterwards, so I switched to a
textfield for the console input.  I then fixed an edge of map issue I
was having with the directions in the console textfield.

Prototype is an easy organizer, especially when adding extra items to a
list.  However, it just extended my code unnecessarily because I was
forced to extend other areas.